add test machine, when replace remote kernel, you need first check this one

+++ script/compile_install_same_arch.sh
@@ -6,28 +6,42 @@ echo $3 #kernel dir
 echo $4 #def config
 echo $5 #threads
 echo $6 #password
+echo $7 #how many times compile
+echo $8 #tot times
 
-sudo yum install expect -y
-sudo apt install expect -y
+res=`expect -v |grep bash`
+if [[ ! -z "$res" ]]
+then
+	sudo yum install expect -y
+	sudo apt install expect -y
+fi
+
 pwd_script=`pwd`
-cd $3
-make $4
-make -j$5 LOCALVERSION=-pqy
-rm -rf $3/mod
-mkdir $3/mod
-make modules_install INSTALL_MOD_PATH=$3/mod -j32
-cd $3/mod/lib/modules/
-tar -cvzf $3/mod.tar.gz 4.19.190-pqy
+
+if [[ "$7" -eq 1 ]];
+then
+	cd $3
+	make $4
+	make -j$5 LOCALVERSION=-pqy
+	rm -rf $3/mod
+	mkdir $3/mod
+	make modules_install INSTALL_MOD_PATH=$3/mod -j32
+	cd $3/mod/lib/modules/
+	tar -cvzf $3/mod.tar.gz 4.19.190-pqy
+fi
 
 cd
 expect << __EOF
 set timeout 30
-spawn scp -o StrictHostKeyChecking=no $3/vmlinuz $3/vmlinux $3/mod.tar.gz ~/kernel_debug_tools/script/startup.sh $1@$2:~
+spawn scp -o StrictHostKeyChecking=no $3/vmlinuz $3/vmlinux $3/mod.tar.gz $HOME/kernel_debug_tools/script/startup.sh $1@$2:~
 expect "*password:"
 send "$6\r"
 expect of
 __EOF
 
-rm $3/mod $3/mod.tar.gz -rf
+if [[ "$7" -eq "$8" ]]
+then
+	rm $3/mod $3/mod.tar.gz -rf
+fi
 cd $pwd_script
 ./install_same_arch $1 $2 $6
